+++ android/src/main/kotlin/com/apphud/fluttersdk/handlers/UserPropertiesHandler.kt
@@ -6,37 +6,59 @@ import com.apphud.sdk.ApphudUserPropertyKey
 import io.flutter.plugin.common.MethodChannel
 
 
-class UserPropertiesHandler(override val routes: List<String>, val context: Context) : Handler {
-
-    override fun tryToHandle(method: String, args: Map<String, Any>?, result: MethodChannel.Result) {
+class UserPropertiesHandler(
+    override val routes: List<String>,
+    val context: Context,
+    handleOnMainThreadP: HandleOnMainThread
+) : Handler {
+    private var handleOnMainThread = handleOnMainThreadP
+    override fun tryToHandle(
+        method: String,
+        args: Map<String, Any>?,
+        result: MethodChannel.Result
+    ) {
         when (method) {
             UserPropertiesRoutes.setUserProperty.name -> SetUserPropertyParser(result).parse(args) { key, value, setOnce ->
                 setUserProperty(key, value, setOnce, result)
             }
 
-            UserPropertiesRoutes.incrementUserProperty.name -> IncrementUserPropertyParser(result).parse(args) { key, by ->
+            UserPropertiesRoutes.incrementUserProperty.name -> IncrementUserPropertyParser(result).parse(
+                args
+            ) { key, by ->
                 incrementUserProperty(key, by, result)
             }
         }
     }
 
-    private fun setUserProperty(key: ApphudUserPropertyKey, value: Any?, setOnce: Boolean, result: MethodChannel.Result) {
+    private fun setUserProperty(
+        key: ApphudUserPropertyKey,
+        value: Any?,
+        setOnce: Boolean,
+        result: MethodChannel.Result
+    ) {
         Apphud.setUserProperty(key = key, value = value, setOnce = setOnce)
-        result.success(null)
+        handleOnMainThread { result.success(null) }
     }
 
-    private fun incrementUserProperty(key: ApphudUserPropertyKey, by: Any, result: MethodChannel.Result) {
+    private fun incrementUserProperty(
+        key: ApphudUserPropertyKey,
+        by: Any,
+        result: MethodChannel.Result
+    ) {
         Apphud.incrementUserProperty(key = key, by = by)
-        result.success(null)
+        handleOnMainThread { result.success(null) }
     }
 
     class SetUserPropertyParser(val result: MethodChannel.Result) : ApphudUserPropertyKeyParser() {
 
-        fun parse(args: Map<String, Any>?, callback: (key: ApphudUserPropertyKey, value: Any?, setOnce: Boolean) -> Unit) {
+        fun parse(
+            args: Map<String, Any>?,
+            callback: (key: ApphudUserPropertyKey, value: Any?, setOnce: Boolean) -> Unit
+        ) {
             try {
                 args ?: throw IllegalArgumentException("key is required argument")
                 val keyString = args["key"] as? String
-                        ?: throw IllegalArgumentException("key is required argument")
+                    ?: throw IllegalArgumentException("key is required argument")
 
                 val key = parseApphudUserPropertyKey(keyString)
 
@@ -51,18 +73,22 @@ class UserPropertiesHandler(override val routes: List<String>, val context: Cont
         }
     }
 
-    class IncrementUserPropertyParser(val result: MethodChannel.Result) : ApphudUserPropertyKeyParser() {
+    class IncrementUserPropertyParser(val result: MethodChannel.Result) :
+        ApphudUserPropertyKeyParser() {
 
-        fun parse(args: Map<String, Any>?, callback: (key: ApphudUserPropertyKey, by: Any) -> Unit) {
+        fun parse(
+            args: Map<String, Any>?,
+            callback: (key: ApphudUserPropertyKey, by: Any) -> Unit
+        ) {
             try {
                 args ?: throw IllegalArgumentException("key is required argument")
                 val keyString = args["key"] as? String
-                        ?: throw IllegalArgumentException("key is required argument")
+                    ?: throw IllegalArgumentException("key is required argument")
 
                 val key = parseApphudUserPropertyKey(keyString)
 
                 val by = args["by"] as? Any
-                        ?: throw IllegalArgumentException("by is required argument")
+                    ?: throw IllegalArgumentException("by is required argument")
 
                 callback(key, by)
             } catch (e: IllegalArgumentException) {