+++ lib/models/apphud_models/apphud_attribution_provider.dart
@@ -1,6 +1,7 @@
 enum ApphudAttributionProvider {
   appsFlyer,
   adjust,
+  // remove this, only appleAdsAttribution is needed. iAd is no longer supported by Apple
   @Deprecated('Attribution for versions 14.2 or lower, iAd framework')
   appleSearchAds,
   facebook,
@@ -15,10 +16,13 @@ extension ApphudAttributionProviderConverter on ApphudAttributionProvider {
         return 'appsFlyer';
       case ApphudAttributionProvider.adjust:
         return 'adjust';
+        // TODO: remove this, only appleAdsAttribution is needed. iAd is no longer supported by Apple
       case ApphudAttributionProvider.appleSearchAds:
         return 'appleSearchAds';
+        // remove facebook
       case ApphudAttributionProvider.facebook:
         return 'facebook';
+        // add firebase to Apphud.dart method description
       case ApphudAttributionProvider.firebase:
         return 'firebase';
       case ApphudAttributionProvider.appleAdsAttribution: