+++ lib/models/apphud_models/apphud_paywall.g.dart
@@ -10,7 +10,6 @@ ApphudPaywall _$ApphudPaywallFromJson(Map json) => ApphudPaywall(
       identifier: json['identifier'] as String,
       isDefault: json['isDefault'] as bool,
       experimentName: json['experimentName'] as String?,
-      variationName: json['variationName'] as String?,
       fromPaywall: json['fromPaywall'] as String?,
       json: (json['json'] as Map?)?.map(
         (k, e) => MapEntry(k as String, e),
@@ -25,7 +24,6 @@ Map<String, dynamic> _$ApphudPaywallToJson(ApphudPaywall instance) =>
       'identifier': instance.identifier,
       'isDefault': instance.isDefault,
       'experimentName': instance.experimentName,
-      'variationName': instance.variationName,
       'fromPaywall': instance.fromPaywall,
       'json': instance.json,
       'products': instance.products,